fix ema for period of one with incorrect timestamps and separate tests
